Cambio en Recursos

+++ funciones.h
@@ -17,30 +17,30 @@ float validarFlotante(int i, int opcFlotante, int opcInstrucciones);
 
 void buscarReemplazarEliminar(char nombresProductos[NUM_PRODUCTOS][NUM_CARACTERES_NOMBRES], int cantidadProductos,
                               int reemplazarEliminar, int *eliminados,
-                              float tiempoUnidad[NUM_PRODUCTOS], int recursosUnidad[NUM_PRODUCTOS],
+                              float tiempoUnidad[NUM_PRODUCTOS], float recursosUnidad[NUM_PRODUCTOS],
                               int cantDemandada[NUM_PRODUCTOS]);
 
-void tiempoRecursos(float *tiempoUnidad, int *recursosUnidad, char nombresProductos[NUM_CARACTERES_NOMBRES], int i);
+void tiempoRecursos(float *tiempoUnidad, float *recursosUnidad, char nombresProductos[NUM_CARACTERES_NOMBRES], int i);
 
 
 
 void menu(char nombresProductos[NUM_PRODUCTOS][NUM_CARACTERES_NOMBRES], int cantOpcsMenu,
           int *cantProductos, int *continuar,
-          float tiempoUnidad[NUM_PRODUCTOS], int recursosUnidad[NUM_PRODUCTOS],
-          float *tiempoDisponible, int *recursosDisponibles,
+          float tiempoUnidad[NUM_PRODUCTOS], float recursosUnidad[NUM_PRODUCTOS],
+          float *tiempoDisponible, float *recursosDisponibles,
           int cantDemandada[NUM_PRODUCTOS]);
 
-void tiempoRecursosRequeridos(float tiempoUnidad[NUM_PRODUCTOS], int recursosUnidad[NUM_PRODUCTOS], int cantDemandada[NUM_PRODUCTOS],
+void tiempoRecursosRequeridos(float tiempoUnidad[NUM_PRODUCTOS], float recursosUnidad[NUM_PRODUCTOS], int cantDemandada[NUM_PRODUCTOS],
                               int cantProductos,
-                              float tiempoDisponible, int recursosDisponibles);
+                              float tiempoDisponible, float recursosDisponibles);
 
 void imprimirLineas();
 
 
 
-void recomendarProduccion(float tiempoUnidad[NUM_PRODUCTOS], int recursosUnidad[NUM_PRODUCTOS], int cantDemandada[NUM_PRODUCTOS],
+void recomendarProduccion(float tiempoUnidad[NUM_PRODUCTOS], float recursosUnidad[NUM_PRODUCTOS], int cantDemandada[NUM_PRODUCTOS],
                           int cantProductos,
-                          float tiempoDisponible, int recursosDisponibles);
+                          float tiempoDisponible, float recursosDisponibles);
